# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Gestão de Projetos API",
    "description": "API para gerenciar o ciclo de vida de projetos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Lista Projetos com Paginação",
        "description": "Retorna uma lista paginada de todos os projetos do sistema. Permite filtrar por status do projeto. Requer autenticação de qualquer usuário válido.",
        "operationId": "read_projects_api_projects__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por status do projeto",
              "title": "Status"
            },
            "description": "Filtrar por status do projeto"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Número da página",
              "default": 1,
              "title": "Page"
            },
            "description": "Número da página"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "exclusiveMinimum": 0,
              "description": "Tamanho da página",
              "default": 20,
              "title": "Size"
            },
            "description": "Tamanho da página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Cria um Novo Projeto",
        "description": "Cria um novo projeto no sistema com os dados fornecidos. Registra automaticamente um log de auditoria da criação. Requer permissão de MANAGER ou ADMIN.",
        "operationId": "create_project_api_projects__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Busca Projeto por ID",
        "description": "Retorna os detalhes completos de um projeto específico pelo seu ID único. Requer autenticação de qualquer usuário válido.",
        "operationId": "get_project_api_projects__project_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{p_id}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Atualiza um Projeto Existente",
        "description": "Atualiza os dados de um projeto existente. Envia notificações automáticas quando há mudança de Project Manager ou Technical Lead. Registra log de auditoria da atualização. Requer permissão de MANAGER ou ADMIN.",
        "operationId": "update_project_api_projects__p_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "p_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "P Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Exclui um Projeto",
        "description": "Remove permanentemente um projeto do sistema. Registra log de auditoria da exclusão. Requer permissão de ADMIN apenas.",
        "operationId": "delete_project_api_projects__p_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "p_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "P Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/advance-phase": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Avança Fase do Projeto",
        "description": "Avança o projeto para a próxima fase do workflow, validando os quality gates necessários. Registra log de auditoria do avanço. Requer permissão de MANAGER ou ADMIN.",
        "operationId": "advance_project_phase_api_projects__project_id__advance_phase_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ProjectRead_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Número total de itens disponíveis na consulta",
            "example": 150
          },
          "page": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Page",
            "description": "Número da página atual (baseado em 1)",
            "example": 2
          },
          "size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size",
            "description": "Número de itens por página",
            "example": 20
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Número total de páginas disponíveis",
            "example": 8
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Lista de itens para a página atual"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "size",
          "pages",
          "items"
        ],
        "title": "PaginatedResponse[ProjectRead]"
      },
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nome descritivo do projeto",
            "example": "Implementação Microsoft 365"
          },
          "client": {
            "type": "string",
            "title": "Client",
            "description": "Nome do cliente para o qual o projeto é executado",
            "example": "Empresa ABC Ltda"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "Data de início do projeto",
            "example": "2025-01-15"
          },
          "estimatedEndDate": {
            "type": "string",
            "format": "date",
            "title": "Estimatedenddate",
            "description": "Data estimada para conclusão do projeto",
            "example": "2025-06-30"
          },
          "orderValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordervalue",
            "description": "Valor do contrato/ordem de serviço",
            "example": "R$ 150.000,00"
          },
          "proposal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal",
            "description": "Número ou referência da proposta comercial",
            "example": "PROP-2024-001"
          },
          "pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct",
            "description": "Número do PCT (Plano de Controle de Trabalho)",
            "example": "PCT-2024-MS365-001"
          },
          "phase": {
            "$ref": "#/components/schemas/ProjectPhase",
            "description": "Fase atual do projeto no ciclo de vida",
            "default": "inception",
            "example": "INCEPTION"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Status operacional do projeto",
            "default": "active",
            "example": "ACTIVE"
          },
          "project_manager_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Manager Id",
            "description": "ID do gerente de projeto responsável",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "technical_lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Lead Id",
            "description": "ID do líder técnico do projeto",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          }
        },
        "type": "object",
        "required": [
          "name",
          "client",
          "startDate",
          "estimatedEndDate"
        ],
        "title": "ProjectCreate"
      },
      "ProjectPhase": {
        "type": "string",
        "enum": [
          "inception",
          "definition",
          "built",
          "deploy",
          "close"
        ],
        "title": "ProjectPhase"
      },
      "ProjectRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nome descritivo do projeto",
            "example": "Implementação Microsoft 365"
          },
          "client": {
            "type": "string",
            "title": "Client",
            "description": "Nome do cliente para o qual o projeto é executado",
            "example": "Empresa ABC Ltda"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "Data de início do projeto",
            "example": "2025-01-15"
          },
          "estimatedEndDate": {
            "type": "string",
            "format": "date",
            "title": "Estimatedenddate",
            "description": "Data estimada para conclusão do projeto",
            "example": "2025-06-30"
          },
          "orderValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordervalue",
            "description": "Valor do contrato/ordem de serviço",
            "example": "R$ 150.000,00"
          },
          "proposal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal",
            "description": "Número ou referência da proposta comercial",
            "example": "PROP-2024-001"
          },
          "pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct",
            "description": "Número do PCT (Plano de Controle de Trabalho)",
            "example": "PCT-2024-MS365-001"
          },
          "phase": {
            "$ref": "#/components/schemas/ProjectPhase",
            "description": "Fase atual do projeto no ciclo de vida",
            "default": "inception",
            "example": "INCEPTION"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Status operacional do projeto",
            "default": "active",
            "example": "ACTIVE"
          },
          "project_manager_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Manager Id",
            "description": "ID do gerente de projeto responsável",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "technical_lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Lead Id",
            "description": "ID do líder técnico do projeto",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único do projeto",
            "example": "550e8400-e29b-41d4-a716-446655440003"
          },
          "project_manager": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInProject"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados do gerente de projeto"
          },
          "technical_lead": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInProject"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados do líder técnico"
          }
        },
        "type": "object",
        "required": [
          "name",
          "client",
          "startDate",
          "estimatedEndDate",
          "id"
        ],
        "title": "ProjectRead"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "active",
          "hold",
          "completed",
          "cancelled"
        ],
        "title": "ProjectStatus"
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Novo nome do projeto",
            "example": "Implementação Microsoft 365 - Fase 2"
          },
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client",
            "description": "Novo nome do cliente",
            "example": "Empresa XYZ Ltda"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate",
            "description": "Nova data de início",
            "example": "2025-02-01"
          },
          "estimatedEndDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedenddate",
            "description": "Nova data estimada de conclusão",
            "example": "2025-08-31"
          },
          "orderValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordervalue",
            "description": "Novo valor do contrato",
            "example": "R$ 200.000,00"
          },
          "proposal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal",
            "description": "Nova referência da proposta",
            "example": "PROP-2024-002"
          },
          "pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct",
            "description": "Novo número do PCT",
            "example": "PCT-2024-MS365-002"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectPhase"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nova fase do projeto",
            "example": "ELABORATION"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Novo status do projeto",
            "example": "HOLD"
          },
          "project_manager_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Manager Id",
            "description": "Novo ID do gerente de projeto",
            "example": "550e8400-e29b-41d4-a716-446655440004"
          },
          "technical_lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Lead Id",
            "description": "Novo ID do líder técnico",
            "example": "550e8400-e29b-41d4-a716-446655440005"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "UserInProject": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único do usuário",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Endereço de email do usuário",
            "example": "joao.silva@empresa.com"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserInProject"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}